# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CRM module for Uralintern",
    "description": "API documentation for Uralintern CRM Module.\n\n**Контроллеры:**\n- **auth-controller**: Отвечает за регистрацию и авторизацию пользователей, включая регистрацию с реферальной ссылкой и обновление токенов.\n- **user-info-controller**: Управляет пользовательскими данными, включая обновление профиля, получение реферальных ссылок и список руководителей.\n- **event-controller**: Обеспечивает управление мероприятиями, включая их создание, обновление, скрытие, удаление, а также доступ к информации о мероприятиях.\n- **message-controller**: Управляет шаблонами сообщений для мероприятий, включая их получение и обновление.\n- **event-student-controller**: Обеспечивает управление заявками и участием студентов в мероприятиях, включая отправку заявок, их принятие/отклонение и управление студентами на мероприятии.\n- **event-curator-controller**: Управляет заявками и участием кураторов в мероприятиях, включая их принятие/отклонение и управление кураторами.\n- **notification-controller**: Управляет уведомлениями пользователей, связанными с их участием в мероприятиях.\n",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost/api",
      "description": "kayakto server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/events-students/test-result": {
      "post": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "saveTestResult",
        "description": "Выгрузка баллов за тест студента. Доступно только администраторам.\nВНИМАНИЕ: этот метод для интеграции с сервисом тестов для uralintern\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentTestResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Баллы успешно выгрузились.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/student-can-send/{studentId}": {
      "get": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "canSendStudent",
        "description": "Если студент отправлял заявку на мероприятие - ему не будет доступна регистрация на мероприятие (вернется false). В обратном случае студент сможет отправить заявку на мероприятие (true).",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/waiting-students": {
      "get": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "getWaitingStudentsInfo",
        "description": "Получение студентов, которые ждут принятия заявки на мероприятие. Доступ: руководитель",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventStudentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/students": {
      "get": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "getStudentsInfo",
        "description": "Получение студентов, которые связаных с данным мероприятием, учитывая все возможные статусы. Доступ: руководитель",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventStudentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/accepted-students": {
      "get": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "getAcceptedStudentsInfo",
        "description": "Получение студентов, которых приняли на мероприятие и их добавили в чат. Доступ: руководитель, администратор.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventStudentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/started-students": {
      "get": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "getStartedStudentsInfo",
        "description": "Получение студентов, которые начали выполнение мероприятия. Доступ: руководитель, администратор.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventStudentInfoDTO"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
    "/events-students/{eventId}/send/{studentId}": {
      "put": {
        "tags": [
          "event-student-controller"
        ],
        "description": "Отправить студенту заявку на регистрацию на мероприятие. Доступ - студент.",
        "operationId": "sendStudentToEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "идентификатор мероприятия",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "идентификатор студента",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/reject/{studentId}": {
      "put": {
        "tags": [
          "event-student-controller"
        ],
        "description": "Отклонить заявку студента на мероприятие. Доступ- руководитель.",
        "operationId": "rejectStudentRequest",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "идентификатор мероприятия",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "идентификатор студента",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/accept/{studentId}": {
      "put": {
        "tags": [
          "event-student-controller"
        ],
        "description": "Принять заявку студента на мероприятие. Доступ - руководитель.",
        "operationId": "acceptStudentRequest",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "идентификатор мероприятия",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "идентификатор студента",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/events-students/{eventId}/delete/{studentId}": {
      "delete": {
        "tags": [
          "event-student-controller"
        ],
        "operationId": "deleteStudentFromEvent",
        "description": "Удаление студента с мероприятия. Доступ - руководитель.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "BadRequestResponse": {
        "description": "Bad Request, запрос неверный со стороны клиента",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid argument"
                }
              }
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "Forbidden, отказано в доступе",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Access denied"
                }
              }
            }
          }
        }
      },
      "NotFoundResponse": {
        "description": "Not Found, такого элемента не существует",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Event with id 123 not found"
                }
              }
            }
          }
        }
      },
      "InternalServerErrorResponse": {
        "description": "Internal Server Error, запрос вызвал ошибку на сервере",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "StudentTestResultRequest": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "eventId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "passed": {
            "type": "boolean",
            "example": true
          },
          "score": {
            "type": "integer",
            "format": "int64",
            "example": 100
          }
        }
      },
...
      "EventStudentInfoDTO": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "studentId": {
            "type": "integer",
            "format": "int64",
            "example": 3
          },
          "statusRequest": {
            "type": "string",
            "enum": [
              "SENT_PERSONAL_INFO",
              "REJECTED_FROM_EVENT",
              "ADDED_IN_CHAT",
              "STARTED_EVENT",
              "ENDED_EVENT",
              "DELETED_FROM_EVENT"
            ],
            "example": "SENT_PERSONAL_INFO"
          },
          "firstName": {
            "type": "string",
            "example": "Николай"
          },
          "lastName": {
            "type": "string",
            "example": "Алексеев"
          },
          "surname": {
            "type": "string",
            "example": "Михайлович"
          },
          "competencies": {
            "type": "string",
            "example": "Spring Boot, Gradle, git, Java, Docker"
          },
          "telegramUrl": {
            "type": "string",
            "example": "https://t.me/kayakto"
          },
          "vkUrl": {
            "type": "string",
            "example": "https://vk.com/kayakto"
          },
          "curatorFirstName": {
            "type": "string",
            "example": "Николай"
          },
          "curatorLastName": {
            "type": "string",
            "example": "Путин"
          },
          "curatorSurname": {
            "type": "string",
            "example": "Евгеньевич"
          }
        }
      },
...
      "MessageResponseDTO": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}